+++ Modules/Intent.Modules.Metadata.RDBMS/Api/Extensions/AttributeModelStereotypeExtensions.cs
@@ -434,6 +434,49 @@ namespace Intent.Metadata.RDBMS.Api
                 return _stereotype.GetProperty<bool>("IsUnique");
             }
 
+            public SortDirectionOptions SortDirection()
+            {
+                return new SortDirectionOptions(_stereotype.GetProperty<string>("Sort Direction"));
+            }
+
+            public class SortDirectionOptions
+            {
+                public readonly string Value;
+
+                public SortDirectionOptions(string value)
+                {
+                    Value = value;
+                }
+
+                public SortDirectionOptionsEnum AsEnum()
+                {
+                    switch (Value)
+                    {
+                        case "Ascending":
+                            return SortDirectionOptionsEnum.Ascending;
+                        case "Descending":
+                            return SortDirectionOptionsEnum.Descending;
+                        default:
+                            throw new ArgumentOutOfRangeException();
+                    }
+                }
+
+                public bool IsAscending()
+                {
+                    return Value == "Ascending";
+                }
+                public bool IsDescending()
+                {
+                    return Value == "Descending";
+                }
+            }
+
+            public enum SortDirectionOptionsEnum
+            {
+                Ascending,
+                Descending
+            }
+
         }
 
         public class PrimaryKey