+++ Modules/Intent.Modules.Metadata.RDBMS/Api/Indexes/SortDirection.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Intent.Modules.Metadata.RDBMS.Api.Indexes
+{
+    public enum SortDirection
+    {
+        Ascending,
+        Descending
+    }
+}